requirements.txt
asgiref==3.3.4
atomicwrites==1.4.0
attrs==21.2.0
colorama==0.4.4
cycler==0.10.0
Django==3.2.3
django-crispy-forms==1.11.2
django-jquery==3.1.0
gunicorn==20.1.0
importlib-metadata==4.0.1
iniconfig==1.1.1
kiwisolver==1.3.1
matplotlib==3.4.2
mysqlclient==2.0.3
numpy==1.20.3
packaging==20.9
pandas==1.2.4
Pillow==8.2.0
pluggy==0.13.1
py==1.10.0
pyparsing==2.4.7
pytest==6.2.4
pytest-django==4.3.0
python-dateutil==2.8.1
pytz==2021.1
six==1.16.0
sqlparse==0.4.1
toml==0.10.2
typing-extensions==3.10.0.0
whitenoise==5.2.0
zipp==3.4.1
